requirements.txt
agent-framework>=1.0.0b251120
pydantic>=2.0.0
typing-extensions>=4.0.0
streamlit>=1.28.0
nest-asyncio>=1.5.6
anthropic>=0.18.0
python-dotenv>=1.0.0

